refactor(domain): extract topic auto-creation into reusable service
Consolidate duplicated topic auto-creation logic from CreateItemWorkflow
and EditItemWorkflow into a shared TopicAutoCreationService.

This removes ~220 lines of duplication while maintaining the same
behavior and deferred-persistence pattern for topic creation.
