+++ src/domain/workflows/create_item.ts
@@ -5,7 +5,6 @@ import {
   AliasSlug,
   CalendarDay,
   createItemIcon,
-  createPermanentPlacement,
   DateTime,
   Duration,
   isCalendarDay,
@@ -24,7 +23,13 @@ import { RepositoryError } from "../repositories/repository_error.ts";
 import { RankService } from "../services/rank_service.ts";
 import { IdGenerationService } from "../services/id_generation_service.ts";
 import { AliasAutoGenerator } from "../services/alias_auto_generator.ts";
-import { Alias, createAlias } from "../models/alias.ts";
+import { createAlias } from "../models/alias.ts";
+import {
+  buildTopicItem,
+  persistPreparedTopic,
+  PreparedTopic,
+  TopicBuildError,
+} from "../services/topic_auto_creation_service.ts";
 
 export type CreateItemInput = Readonly<{
   title: string;
@@ -88,122 +93,11 @@ const repositoryFailure = (error: RepositoryError): CreateItemRepositoryError =>
   error,
 });
 
-/**
- * A prepared topic ready to be persisted. Contains the Item and Alias that
- * will be saved together when the main workflow validation passes.
- */
-type PreparedTopic = Readonly<{
-  item: Item;
-  alias: Alias;
-  slug: AliasSlug;
-}>;
-
-/**
- * Builds a topic item and alias without persisting them.
- * Used to defer persistence until after all validation passes.
- */
-const buildTopicItem = async (
-  aliasSlug: AliasSlug,
-  createdAt: DateTime,
-  deps: CreateItemDependencies,
-): Promise<Result<PreparedTopic, CreateItemError>> => {
-  // Generate ID for the new topic
-  const idResult = deps.idGenerationService.generateId();
-  if (idResult.type === "error") {
-    return Result.error(invalidInput(
-      idResult.error.issues.map((issue) =>
-        createValidationIssue(issue.message, {
-          code: issue.code,
-          path: ["topic", "id", ...issue.path],
-        })
-      ),
-    ));
+const topicBuildErrorToCreateItemError = (error: TopicBuildError): CreateItemError => {
+  if (error.kind === "validation") {
+    return invalidInput(error.issues);
   }
-  const topicId = idResult.value;
-
-  // Use the alias as the title
-  const titleResult = itemTitleFromString(aliasSlug.toString());
-  if (titleResult.type === "error") {
-    return Result.error(invalidInput(
-      titleResult.error.issues.map((issue) =>
-        createValidationIssue(issue.message, {
-          code: issue.code,
-          path: ["topic", "title", ...issue.path],
-        })
-      ),
-    ));
-  }
-  const title = titleResult.value;
-
-  // Get rank for permanent placement
-  const permanentPlacement = createPermanentPlacement();
-  const siblingsResult = await deps.itemRepository.listByPlacement({
-    kind: "single",
-    at: permanentPlacement,
-  });
-  if (siblingsResult.type === "error") {
-    return Result.error(repositoryFailure(siblingsResult.error));
-  }
-  const existingRanks = siblingsResult.value.map((item) => item.data.rank);
-  const rankResult = deps.rankService.tailRank(existingRanks);
-  if (rankResult.type === "error") {
-    return Result.error(invalidInput(
-      rankResult.error.issues.map((issue) =>
-        createValidationIssue(issue.message, {
-          code: issue.code,
-          path: ["topic", "rank", ...issue.path],
-        })
-      ),
-    ));
-  }
-
-  // Create the topic item (not persisted yet)
-  const topicItem = createItem({
-    id: topicId,
-    title,
-    icon: createItemIcon("topic"),
-    status: itemStatusOpen(),
-    placement: permanentPlacement,
-    rank: rankResult.value,
-    createdAt,
-    updatedAt: createdAt,
-    alias: aliasSlug,
-  });
-
-  // Create the alias model (not persisted yet)
-  const aliasModel = createAlias({
-    slug: aliasSlug,
-    itemId: topicId,
-    createdAt,
-  });
-
-  return Result.ok({
-    item: topicItem,
-    alias: aliasModel,
-    slug: aliasSlug,
-  });
-};
-
-/**
- * Persists a prepared topic (item and alias) to the repositories.
- */
-const persistPreparedTopic = async (
-  prepared: PreparedTopic,
-  deps: CreateItemDependencies,
-): Promise<Result<void, CreateItemError>> => {
-  // Save the topic item
-  const saveResult = await deps.itemRepository.save(prepared.item);
-  if (saveResult.type === "error") {
-    return Result.error(repositoryFailure(saveResult.error));
-  }
-
-  // Save the alias
-  const aliasSaveResult = await deps.aliasRepository.save(prepared.alias);
-  if (aliasSaveResult.type === "error") {
-    return Result.error(repositoryFailure(aliasSaveResult.error));
-  }
-
-  return Result.ok(undefined);
+  return repositoryFailure(error.error);
 };
 
 /**
@@ -349,7 +243,7 @@ export const CreateItemWorkflow = {
             deps,
           );
           if (buildResult.type === "error") {
-            return Result.error(buildResult.error);
+            return Result.error(topicBuildErrorToCreateItemError(buildResult.error));
           }
           projectId = buildResult.value.item.data.id;
           pendingTopics.push(buildResult.value);
@@ -412,7 +306,7 @@ export const CreateItemWorkflow = {
                 deps,
               );
               if (buildResult.type === "error") {
-                return Result.error(buildResult.error);
+                return Result.error(topicBuildErrorToCreateItemError(buildResult.error));
               }
               contextIds.push(buildResult.value.item.data.id);
               processedAliases.set(aliasKey, buildResult.value.item.data.id);
@@ -594,7 +488,7 @@ export const CreateItemWorkflow = {
     for (const prepared of pendingTopics) {
       const persistResult = await persistPreparedTopic(prepared, deps);
       if (persistResult.type === "error") {
-        return Result.error(persistResult.error);
+        return Result.error(repositoryFailure(persistResult.error));
       }
     }
