+++ src/domain/workflows/edit_item.ts
@@ -4,17 +4,13 @@ import {
   createValidationIssue,
   ValidationError,
 } from "../../shared/errors.ts";
-import { createItem, Item } from "../models/item.ts";
-import { Alias, createAlias } from "../models/alias.ts";
+import { Item } from "../models/item.ts";
+import { createAlias } from "../models/alias.ts";
 import {
   AliasSlug,
-  createItemIcon,
-  createPermanentPlacement,
   DateTime,
   Duration,
   ItemId,
-  itemStatusOpen,
-  itemTitleFromString,
   parseAliasSlug,
   parseDateTime,
   parseDuration,
@@ -28,6 +24,12 @@ import { AliasRepository } from "../repositories/alias_repository.ts";
 import { RepositoryError } from "../repositories/repository_error.ts";
 import { RankService } from "../services/rank_service.ts";
 import { IdGenerationService } from "../services/id_generation_service.ts";
+import {
+  buildTopicItem,
+  persistPreparedTopic,
+  PreparedTopic,
+  TopicBuildError,
+} from "../services/topic_auto_creation_service.ts";
 
 export type EditItemInput = Readonly<{
   itemLocator: string;
@@ -65,131 +67,11 @@ export type EditItemResult = Readonly<{
   createdTopics: ReadonlyArray<AliasSlug>;
 }>;
 
-/**
- * A prepared topic ready to be persisted. Contains the Item and Alias that
- * will be saved together when the main workflow validation passes.
- */
-type PreparedTopic = Readonly<{
-  item: Item;
-  alias: Alias;
-  slug: AliasSlug;
-}>;
-
-/**
- * Builds a topic item and alias without persisting them.
- * Used to defer persistence until after all validation passes.
- */
-const buildTopicItem = async (
-  aliasSlug: AliasSlug,
-  createdAt: DateTime,
-  deps: EditItemDependencies,
-): Promise<Result<PreparedTopic, EditItemError>> => {
-  // Generate ID for the new topic
-  const idResult = deps.idGenerationService.generateId();
-  if (idResult.type === "error") {
-    return Result.error(
-      createValidationError(
-        "EditItem",
-        idResult.error.issues.map((issue) =>
-          createValidationIssue(issue.message, {
-            code: issue.code,
-            path: ["topic", "id", ...issue.path],
-          })
-        ),
-      ),
-    );
+const topicBuildErrorToEditItemError = (error: TopicBuildError): EditItemError => {
+  if (error.kind === "validation") {
+    return createValidationError("EditItem", error.issues);
   }
-  const topicId = idResult.value;
-
-  // Use the alias as the title
-  const titleResult = itemTitleFromString(aliasSlug.toString());
-  if (titleResult.type === "error") {
-    return Result.error(
-      createValidationError(
-        "EditItem",
-        titleResult.error.issues.map((issue) =>
-          createValidationIssue(issue.message, {
-            code: issue.code,
-            path: ["topic", "title", ...issue.path],
-          })
-        ),
-      ),
-    );
-  }
-  const title = titleResult.value;
-
-  // Get rank for permanent placement
-  const permanentPlacement = createPermanentPlacement();
-  const siblingsResult = await deps.itemRepository.listByPlacement({
-    kind: "single",
-    at: permanentPlacement,
-  });
-  if (siblingsResult.type === "error") {
-    return Result.error(siblingsResult.error);
-  }
-  const existingRanks = siblingsResult.value.map((item) => item.data.rank);
-  const rankResult = deps.rankService.tailRank(existingRanks);
-  if (rankResult.type === "error") {
-    return Result.error(
-      createValidationError(
-        "EditItem",
-        rankResult.error.issues.map((issue) =>
-          createValidationIssue(issue.message, {
-            code: issue.code,
-            path: ["topic", "rank", ...issue.path],
-          })
-        ),
-      ),
-    );
-  }
-
-  // Create the topic item (not persisted yet)
-  const topicItem = createItem({
-    id: topicId,
-    title,
-    icon: createItemIcon("topic"),
-    status: itemStatusOpen(),
-    placement: permanentPlacement,
-    rank: rankResult.value,
-    createdAt,
-    updatedAt: createdAt,
-    alias: aliasSlug,
-  });
-
-  // Create the alias model (not persisted yet)
-  const aliasModel = createAlias({
-    slug: aliasSlug,
-    itemId: topicId,
-    createdAt,
-  });
-
-  return Result.ok({
-    item: topicItem,
-    alias: aliasModel,
-    slug: aliasSlug,
-  });
-};
-
-/**
- * Persists a prepared topic (item and alias) to the repositories.
- */
-const persistPreparedTopic = async (
-  prepared: PreparedTopic,
-  deps: EditItemDependencies,
-): Promise<Result<void, EditItemError>> => {
-  // Save the topic item
-  const saveResult = await deps.itemRepository.save(prepared.item);
-  if (saveResult.type === "error") {
-    return Result.error(saveResult.error);
-  }
-
-  // Save the alias
-  const aliasSaveResult = await deps.aliasRepository.save(prepared.alias);
-  if (aliasSaveResult.type === "error") {
-    return Result.error(aliasSaveResult.error);
-  }
-
-  return Result.ok(undefined);
+  return error.error;
 };
 
 export const EditItemWorkflow = {
@@ -324,7 +206,7 @@ export const EditItemWorkflow = {
               deps,
             );
             if (buildResult.type === "error") {
-              return Result.error(buildResult.error);
+              return Result.error(topicBuildErrorToEditItemError(buildResult.error));
             }
             projectId = buildResult.value.item.data.id;
             pendingTopics.push(buildResult.value);
@@ -387,7 +269,7 @@ export const EditItemWorkflow = {
                   deps,
                 );
                 if (buildResult.type === "error") {
-                  return Result.error(buildResult.error);
+                  return Result.error(topicBuildErrorToEditItemError(buildResult.error));
                 }
                 contextIds.push(buildResult.value.item.data.id);
                 processedAliases.set(aliasKey, buildResult.value.item.data.id);
